Publications: architecture, metrics, vocabulary
### Model architecture
(1) deleted papers that are not widely used in MT, (2) added Luong et al - considered as one of the foundational papers and often mentioned along with Bahdanau et al; generally this is considered general knowledge of the field. Noticed that it's also in the "Attention section", but think that this is a more suitable place for it. Also swapped places the CNN model with Transformer: the latter one is more important (Transformer is essential, CNNs would be good to know) and the timeline is also in this order.

### Metrics
Made categories for metrics, added one-sentence descriptions for each category.
Deleted:
- Och, 2003 - no evaluation metrics proposed
Added:
 - NIST (Doddington, 2002)
 - chrF (Popovic, 2015)
 - BLEURT
 - BertScore
 - MoverScore
P.S. For Embedding-based category, there were a couple more attempts such as e.g. BLEU2VEC, but they didn’t receive any kind of attention - didn’t put them here.

### Open vocabulary
Added:
SentencePiece (Kudo, 2018) - often used in place of BPE